+++ packages/core/src/querier/decorator/index.ts
@@ -1,2 +1,3 @@
 export * from './injectQuerier.js';
+export * from './serialized.js';
 export * from './transactional.js';